# openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    UpsertionDetails:
      description: Details for upserting documents in a collection.
      properties:
        documents:
          anyOf:
          - type: object
          - items:
              type: object
            type: array
          description: The document(s) to upsert.
          title: Documents
        id_field:
          default: _id
          description: The field to use as the document id.
          title: Id Field
          type: string
      required:
      - documents
      title: UpsertionDetails
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: A service for basic infrastructure technology state management.
  title: State Management Service
  version: 1.2.0
openapi: 3.1.0
paths:
  /documents/permissions:
    get:
      description: Returns the configured db permissions list.
      operationId: get_db_permissions
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  type: string
                title: Response Get Db Permissions
                type: array
          description: Successful Response
      summary: Returns the configured db permissions list.
      tags:
      - StateManagementService
      - sms-mongodb
  /documents/{namespace}:
    delete:
      description: No error is raised if the db or collection do not exist.
      operationId: delete_documents
      parameters:
      - description: 'The database and collection to query. Format: db_name.collection'
        in: path
        name: namespace
        required: true
        schema:
          description: 'The database and collection to query. Format: db_name.collection'
          examples:
          - my_test_db.users
          pattern: ^(?P<db>[^.\\/ \"$]{1,64})\.(?P<collection>[^$]{1,255})$
          title: Namespace
          type: string
      - description: Query parameters used to specify the documents affected by the
          request.
        in: query
        name: criteria
        required: false
        schema:
          additionalProperties: true
          description: Query parameters used to specify the documents affected by
            the request.
          examples:
          - '{"name": "John"}'
          - '{"_id": {"$ne": "507f1f77bcf86cd799439011"}}'
          - '{"age": {"$gt": 21}}'
          title: Criteria
          type: object
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Deletes all or some documents in the collection.
      tags:
      - StateManagementService
      - sms-mongodb
    get:
      description: Returns all or some documents from the specified collection.
      operationId: get_documents
      parameters:
      - description: 'The database and collection to query. Format: db_name.collection'
        in: path
        name: namespace
        required: true
        schema:
          description: 'The database and collection to query. Format: db_name.collection'
          examples:
          - my_test_db.users
          pattern: ^(?P<db>[^.\\/ \"$]{1,64})\.(?P<collection>[^$]{1,255})$
          title: Namespace
          type: string
      - description: Query parameters used to specify the documents affected by the
          request.
        in: query
        name: criteria
        required: false
        schema:
          additionalProperties: true
          description: Query parameters used to specify the documents affected by
            the request.
          examples:
          - '{"name": "John"}'
          - '{"_id": {"$ne": "507f1f77bcf86cd799439011"}}'
          - '{"age": {"$gt": 21}}'
          title: Criteria
          type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  type: object
                title: Response Get Documents
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Returns all or some documents from the specified collection.
      tags:
      - StateManagementService
      - sms-mongodb
    put:
      description: Upserts the document(s) provided in the request body in the specified
        collection.
      operationId: upsert_documents
      parameters:
      - description: 'The database and collection to query. Format: db_name.collection'
        in: path
        name: namespace
        required: true
        schema:
          description: 'The database and collection to query. Format: db_name.collection'
          examples:
          - my_test_db.users
          pattern: ^(?P<db>[^.\\/ \"$]{1,64})\.(?P<collection>[^$]{1,255})$
          title: Namespace
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertionDetails'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Upserts the document(s) provided in the request body in thespecified
        collection.
      tags:
      - StateManagementService
      - sms-mongodb
  /health:
    get:
      description: Used to test if this service is alive
      operationId: health
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: health
      tags:
      - StateManagementService
      - health
  /objects/{bucket_id}:
    delete:
      description: Delete all objects in the specified bucket.
      operationId: empty_bucket
      parameters:
      - in: path
        name: bucket_id
        required: true
        schema:
          title: Bucket Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete all objects in the specified bucket.
      tags:
      - StateManagementService
      - sms-s3
    get:
      description: Return a list of the objects currently existing in the S3 object
        storage.
      operationId: get_objects
      parameters:
      - in: path
        name: bucket_id
        required: true
        schema:
          title: Bucket Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  type: string
                title: Response Get Objects
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List all objects in the specified bucket.
      tags:
      - StateManagementService
      - sms-s3
  /objects/{bucket_id}/{object_id}:
    get:
      description: Return a list of the objects currently existing in the S3 object
        storage.
      operationId: check_objects_exists
      parameters:
      - in: path
        name: bucket_id
        required: true
        schema:
          title: Bucket Id
          type: string
      - in: path
        name: object_id
        required: true
        schema:
          title: Object Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Check Objects Exists
                type: boolean
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: See if an object exists in the specified bucket.
      tags:
      - StateManagementService
      - sms-s3
tags:
- name: StateManagementService